perf: reduce dev route compile overhead
