+++ src/app/api/ai/status/route.ts
@@ -1,6 +1,6 @@
-import { NextResponse } from "next/server";
-import { auth } from "@/lib/auth";
-import { getAIStatus } from "@/lib/ai";
+import { NextResponse } from 'next/server';
+import { auth } from '@/lib/auth';
+import { getAIStatus } from '@/lib/ai/config';
 
 export async function GET() {
   const session = await auth();
