+++ src/lib/ai/image-url.ts
@@ -1,0 +1,3 @@
+export function imageUrlFor(generation: { id: string; updatedAt: Date }) {
+  return `/api/ai/wrong-note-image/${generation.id}/file?v=${generation.updatedAt.getTime()}`;
+}
